requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
openai>=2.21.0
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2
